// src/core/types/ThemeOverrides.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ThemeAlpha } from "./ThemeAlpha";
import type { ThemeColors } from "./ThemeColors";
import type { ThemeDimensions } from "./ThemeDimensions";

export type ThemeOverrides = { colors: ThemeColors | null, dimensions: ThemeDimensions | null, alpha: ThemeAlpha | null, };